server/requirements.txt
flask
flask-cors
psycopg2
python-dotenv
pytest
simplejson